% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_country_plot.R
\name{covid_country_plot}
\alias{covid_country_plot}
\title{covid_country_plot}
\usage{
covid_country_plot(country, start_date, end_date)
}
\arguments{
\item{country}{a string of country name that start with capital case}

\item{start_date}{beginning of time interval : "yyyy-mm-dd"}

\item{end_date}{end of time interval : "yyyy-mm-dd"}
}
\value{
a 2-d plot
}
\description{
provides a function to draw 2-d plot for a country death and confirmed covid data, based on Johns Hopkins University's information
}
\examples{
covid_country_plot("Iran","2020-03-24","2020-08-17")
}
