% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_world_map.R
\name{covid_world_map}
\alias{covid_world_map}
\title{covid_world_map}
\usage{
covid_world_map(date, type)
}
\arguments{
\item{date}{a date string like "yyyy-mm-dd"}

\item{type}{should be equal to "death" or "confirmed"}
}
\value{
a world map plot
}
\description{
provides a function to draw Covid-19 world map based on Johns Hopkins University's data
}
\examples{
covid_world_map("2020-03-24","death")
}
