% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-reports.R
\name{get_reports}
\alias{get_reports}
\title{Get the data about COVID-19 from the CSSE repository}
\usage{
get_reports(start_date = Sys.Date() - 1, end_date = Sys.Date() - 1,
  verbose = FALSE, repository = COVID19_repository_dayli)
}
\arguments{
\item{start_date}{Date object YYYY-MM-DD (default to a date before the current date)}

\item{end_date}{Date object YYYY-MM-DD (default to a date before the current date)}

\item{verbose}{should you print file path and readr message? Default to FALSE}

\item{...}{additional args. See below:}
}
\value{
data.frame. It should have the same structure as in here:
https://github.com/CSSEGISandData/COVID-19/blob/master/csse_covid_19_data/csse_covid_19_daily_reports/01-22-2020.csv
}
\description{
Get the data about COVID-19 from the CSSE repository
}
\details{
\itemize{
\item repository: default to the COVID19_repository_daily global variable
\item verbose: should it print out some messages. Default to TRUE
}
}
\examples{
covid_cases = rcovid19::get_reports(Sys.Date() - 1, Sys.Date() - 30)
}
