% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_covid_data.R
\name{update_covid_data}
\alias{update_covid_data}
\title{update_covid_data}
\usage{
update_covid_data()
}
\value{
a data-frame with this variables: country, date, death_num, confirmed_num, recovered_num
}
\description{
provides a function to download the latest Corona Virus data from Johns Hopkins University
}
\examples{
data <- update_covid_data()
}
